cast axis range as double because TH1 axis ranges are also Double_t

+++ util/scriptFiles/plotHead.cc
@@ -3,8 +3,8 @@ struct Plot1DInfoStruct{
     std::string identifier;
     std::string title;
     int nbins;
-    float xmin;
-    float xmax;
+    double xmin;
+    double xmax;
     //std::unique_ptr<TH1> histoptr;
 };
